+++ R/brm_multiple.R
@@ -74,10 +74,14 @@ brm_multiple <- function(formula, data, family = gaussian(), prior = NULL,
                          sparse = NULL, knots = NULL, stanvars = NULL,
                          stan_funs = NULL, recompile = FALSE,
                          combine = TRUE, fit = NA,
-                         seed = NA, file = NULL, ...) {
+                         seed = NA, file = NULL, file_refit = "never", ...) {
   
   combine <- as_one_logical(combine)
+  file_refit <- match.arg(file_refit, c("never", "on_change"))
   if (!is.null(file)) {
+    if(file_refit == "on_change") {
+      stop2("file_refit = 'on_change' is not supported for brm_multiple yet.")
+    }
     # optionally load saved model object
     if (!combine) {
       stop2("Cannot use 'file' if 'combine' is FALSE.")