+++ NEWS.md
@@ -13,6 +13,8 @@ via function `R2D2` to be used in `set_prior`.
 * Extend support for `arma` correlation structures in non-normal families.
 * Extend scope of variables passed via `data2` for use in the 
 evaluation of most model terms.
+* Refit models previously stored on disc only when necessary thanks to
+Martin Modrak. The behavior can be controlled via `file_refit`. (#1058)
 
 ### Other Changes